refactor: useJourneyCreateMutation
