+++ apps/journeys-admin/src/components/OnboardingPanelContent/OnboardingPanelContent.tsx
@@ -13,7 +13,7 @@ import FilePlus1Icon from '@core/shared/ui/icons/FilePlus1'
 import Grid1Icon from '@core/shared/ui/icons/Grid1'
 
 import { GetOnboardingJourneys_onboardingJourneys as OnboardingJourneys } from '../../../__generated__/GetOnboardingJourneys'
-import { useJourneyCreate } from '../../libs/useJourneyCreate/useJourneyCreate'
+import { useJourneyCreateMutation } from '../../libs/useJourneyCreateMutation/useJourneyCreateMutation'
 import { ContainedIconButton } from '../ContainedIconButton'
 import { MediaListItem } from '../MediaListItem'
 import { SidePanelContainer } from '../NewPageWrapper/SidePanelContainer'
@@ -39,7 +39,7 @@ export function OnboardingPanelContent({
   const router = useRouter()
   const { teams } = useFlags()
 
-  const { createJourney, loading } = useJourneyCreate()
+  const { createJourney, loading } = useJourneyCreateMutation()
 
   const templates: OnboardingJourneys[] = []
   onboardingJourneys.forEach((onboardingJourney) => {
