--- apps/journeys-admin/src/libs/useJourneyCreate/index.ts
@@ -1,1 +1,0 @@
-export { useJourneyCreate, CREATE_JOURNEY } from './useJourneyCreate'
