+++ apps/journeys-admin/src/libs/useJourneyCreateMutation/index.ts
@@ -1,0 +1,4 @@
+export {
+  useJourneyCreateMutation,
+  CREATE_JOURNEY
+} from './useJourneyCreateMutation'
